+++ src/main/java/warmup/one/WarmupOne.java
@@ -51,4 +51,17 @@ public class WarmupOne {
         int difference = n - 21;
         return n > 21 ? Math.abs(difference)*2 : Math.abs(difference);
     }
+
+    /**
+     * We have a loud talking parrot. The "hour" parameter is the current hour time in the range
+     * 0..23. We are in trouble if the parrot is talking and the hour is before 7 or after 20. Return
+     * true if we are in trouble.
+     *
+     * @param talking
+     * @param hour
+     * @return
+     */
+    public boolean parrotTrouble(boolean talking, int hour) {
+        return talking && (hour < 7 || hour > 20);
+    }
 }